Popravio autentifikaciju napongit add .

+++ front/src/app/services/AuthService.ts
@@ -2,20 +2,17 @@ import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { HttpHeaders } from '@angular/common/http';
 import { catchError, map } from 'rxjs/operators';
 import { Router } from '@angular/router';
-import { of } from 'rxjs/internal/observable/of';
-import { Observable } from 'rxjs';
 import {ApiService} from "./ApiService";
 import {ConfigService} from "./ConfigService";
 import {UserAppService} from "./UserAppService"
 import {Employee,Credentials} from "../User";
-import { Platform } from '@angular/cdk/platform';
 import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthService {
-  access_token: any;
+  access_token: String | null;
 
   constructor(
     private apiService: ApiService,
@@ -25,9 +22,10 @@ export class AuthService {
     @Inject(PLATFORM_ID) private platformId:Object
   ) {
     if(isPlatformBrowser(this.platformId)){
-      var access_token = localStorage.getItem("jwt");
+      this.access_token = localStorage.getItem("jwt");
     }
   }
+  
 
   
   
@@ -43,11 +41,14 @@ export class AuthService {
     };
     return this.apiService.post(this.config.login_url, JSON.stringify(body), loginHeaders)
       .pipe(map((res) => {
-        console.log(res);
+        
         this.access_token = res.accessToken;
         this.userAppService.getActiveUser();
-        localStorage.setItem("jwt", res.accessToken)
-        window.location.href = "/";
+        if(isPlatformBrowser(this.platformId)){
+          localStorage.setItem("jwt", res.accessToken)
+        }
+        
+        
 
       }));
   }
@@ -70,14 +71,15 @@ export class AuthService {
   logout() {
     this.userAppService.currentUser = null;
     this.access_token = null;
-    localStorage.clear();
+    if(isPlatformBrowser(this.platformId)){localStorage.clear();}
     return this.apiService.post(`${this.config.log_out_url}`, "").subscribe(result => {
       window.location.href = "/";
     });
   }
 
   tokenIsPresent() {
-    return localStorage.getItem("jwt") !== null;
+    if(isPlatformBrowser(this.platformId))
+      return localStorage.getItem("jwt") !== null;
   }
 
   refreshToken() {
